requirements.txt
streamlit
moviepy<2.0
Pillow
numpy
scipy
